requirements.txt
APScheduler==3.6.3
asgiref==3.2.10
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
configparser==5.0.1
cryptography==3.1.1
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.5.0
django-debug-toolbar==3.1.1
django-environ==0.4.5
django-filter==2.4.0
django-oauth-toolkit==1.3.3
django-rest-framework==0.1.0
django-storages==1.10.1
djangorestframework==3.12.1
EasyProcess==0.3
emoji==0.6.0
future==0.18.2
idna==2.10
jsonschema==2.6.0
MeaningCloud-python==2.0.0
numpy==1.19.4
oauthlib==3.1.0
pandas==1.1.4
plyer==1.4.3
pprintpp==0.4.0
protobuf==3.13.0
pycparser==2.20
python-dateutil==2.8.1
python-telegram-bot==13.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.11.13
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tornado==6.0.4
tqdm==4.50.2
tzlocal==2.1
urllib3==1.25.10
webdriverdownloader==1.1.0.3
